[Flutter][Training][Api Chopper] Solved

+++ Dicoding/Flutter/train_api_chopper/lib/model/movie.dart
@@ -24,8 +24,8 @@ class Movie {
   @JsonKey(name: 'release_date') String date;
   @JsonKey(name: 'poster_path') String poster;
   @JsonKey(name: 'backdrop_path') String backdrop;
-  @JsonKey(name: 'vote_average') int average;
-  @JsonKey(name: 'vote_count') int count;
+  @JsonKey(name: 'vote_average') double average;
+  @JsonKey(name: 'vote_count') double count;
 
   Movie({
     required this.id,
